dowload all new versions files from document, add search at documents, move functions to custom_function.py, changes at project
New features in all apps & fix bugs

search function work at all apps / others updates

Bug fixes

Logout implemented!

add username to base.html template

testes for profile & django-global-permissions

first steps for registration

Change core templates directory & update requirements.txt

Tailwind added to categories templates

50% projects with tailwind

Tailwind added to edit project template

tailwind added to project details, first steps home documents

Home documents ended

tailwind added to documents

HTML updated , register, login updated populate updated

Update README.md

added companies & agreements apps

added units app

App compositions , first steps - many bug fixes at directory file system - update populate

Compositions app, new trials - updated populate.py

app composition, paginator added, JS added to new/edit but not work properly

search2 function added to app compositions, fixed separar_valores_sem_espaco function when number and text send

Fixed bug implemented in last commit in separar_valores_sem_espaco function

function separar_valores_sem_espaco updated, search components for select implemented, new_edit template compositions updated

updated search_components function, fixed separar_valores_sem_espaco function

App compositions, JS to add new components, now working

minor changes

JS fill-select adjust to Django form accept new components

Fixed compositions app updates, now the greater code is respected

New updated at app compositions

Implemented new updates at app Compositions (add details.html), change at DB, populate_compositions updated

remove unnecessary output

remove unnecessary output - fix wrong files commited

implemented m2m at composition details, composition template updated

duplicate method implemented at Compositions App

first steps edit composition

app Compositions, listing components using input - TRIAL

App composition - old and new components at same page - check if new components exists not working yet

App compositions, javascript and views.py updated , now inserting and remove components - tech debt->remove all elements not swap iscomposition checkbox to not composition

app compositions new implementations

transfer

transfer

unknown

budget formset factory - trials

budgets functionallity added, fix search for all components at search-components.js

fixed search-components.js , query.trim().length added

Added tailwind local files

insert compositions at budget

repeat category after change selectbox, put fonts and alpine js locally

budget other improvements, add new items and show at screen

budget elements added

Budget application logic working

Solved an issue that caused agreements for the wrong project to be presented upon return

1st trial prices versus compositions

price working, but using only one price

Update to Django 5.1.1
Compositions prices updated (burdened & place added)
ER model pdf updated

Implemented Middleware to require authentication by default that is new in Django 5.1

categories app has big changes, category may have multiples parents

update readme.md , .gitignore & 2 others

rebrand catergories app

added add_or_edit in categories

add categories remodelling

+++ makemake/sites/views.py
@@ -8,26 +8,102 @@ from makemake.sites.forms import SiteForm
 from django.db import IntegrityError
 from django.contrib import messages
 
-@login_required
+from django.http import HttpResponse
+from django.views.generic import View
+
+from django.db.models import Q
+
+import re
+
+from makemake.core.custom_functions import separar_valores_sem_espaco, is_list_empty
+
+
 def delete(request, pk):
-    project = Site.objects.get(pk=pk)
-    try:
-        register_to_delete = get_object_or_404(Site, pk=pk)
-        register_to_delete.delete()
-    except IntegrityError:
-        message = "The register couldn't be deleted!"
+    if request.user.has_perm('global_permissions.entra_em_sites'):
+        project = Site.objects.get(pk=pk)
+        try:
+            register_to_delete = get_object_or_404(Site, pk=pk)
+            register_to_delete.delete()
+        except IntegrityError:
+            message = "The register couldn't be deleted!"
+            messages.info(request, message)
+    else:
+        message = "You cannot have permission to do this!"
         messages.info(request, message)
     return home(request)
     
-@login_required
+
 def home(request):
     items = Site.objects.all()
+    messages.info(request, None)
     return render(request, 'sites/home.html', {'items': items})
 
-@login_required
+
+
+def search(request):
+    # Expressão regular
+    # String de exemplo
+    #string = "   123   /   456  "
+
+    regex = r"^\s*(\d+)\s*\/\s*(\d+)\s*$"
+
+    # Texto de consulta enviado através do formulário
+    input_text = request.GET.get('search', None)
+
+    if input_text == '' or input_text is None:
+        items = Site.objects.all().order_by('name')
+        return render(request, 'sites/home.html', {'items': items})
+
+    # Verifica se a string corresponde ao padrão
+    match = re.match(regex, input_text)
+
+    if match: # Check if pattern is code / year
+        x = int(match.group(1))  # Captura o valor de X
+        y = int(match.group(2))  # Captura o valor de Y
+        # Use Q objects para filtrar seu modelo
+        items = Site.objects.filter(Q(name=x) & Q(name=y)).order_by('name')
+    else:   # Pattern is a free string
+        and_list, or_list = separar_valores_sem_espaco(input_text)
+        if is_list_empty(and_list) and is_list_empty(or_list):
+            items = Site.objects.filter(Q(name__icontains=input_text)).order_by('name')
+        else:
+            ### FUNCAO ABAIXO NÃO ESTA FUNCIONANDO
+            ### PRECISA REFATORAR PARA FICAR MUITO MELHOR
+            string_and = ''
+            string_or = ''
+            pattern = 'Q(name__icontains='
+            for index, value in enumerate(and_list):
+                string_and += pattern + '"' + value + '"' + ')'
+                if index + 1 != len(and_list):
+                    string_and += ' & '
+                
+            for index, value in enumerate(or_list):
+                string_or += pattern + '"' + value + '"' + ')'
+                if index + 1 != len(or_list):
+                    string_or += ' | '
+            final_string = string_and + " | " + string_or
+            if not is_list_empty(and_list) and  is_list_empty(or_list):
+                final_string = string_and
+            elif is_list_empty(and_list) and  not is_list_empty(or_list):
+                final_string = string_or
+            items = Site.objects.filter(eval(final_string)).order_by('name')
+    return render(request, 'sites/home.html', {'items': items})
+
+
+def void(View):
+    def dispatch(self, request, *args, **kwargs):
+            # Coloque aqui sua lógica que não retorna nada para o template
+            # Por exemplo, processamento de dados, atualizações no banco de dados, etc.
+            
+            # Após executar a lógica, você pode simplesmente retornar uma resposta vazia
+            return HttpResponse(status=204)  # 204 significa "No Content"
+
+
 def new(request):
-    #extra_forms = 1  # You can set the initial number of forms here
-    #ProjectBuildingFormSet = formset_factory(ProjectBuildingForm, extra=extra_forms)
+    if not request.user.has_perm('global_permissions.entra_em_sites'):
+        message = "You cannot have permission to do this!"
+        messages.info(request, message)
+        return home(request)
 
     if request.method == 'POST':    # Newly filled form
         form = SiteForm(request.POST or None)
@@ -52,26 +128,29 @@ def new(request):
     context = {'form': form}
     return render(request, 'sites/new_or_edit.html', context)
 
-@login_required
+
 def edit(request, pk=None):
     #extra_forms = 1  # You can set the initial number of forms here
     #ProjectBuildingFormSet = formset_factory(ProjectBuildingForm, extra=extra_forms)
+    if request.user.has_perm('global_permissions.entra_em_sites'):
 
-    if request.method == 'POST':    # Newly filled form
-        form = SiteForm(request.POST or None)
+        if request.method == 'POST':    # Newly filled form
+            form = SiteForm(request.POST or None)
 
-        if form.is_valid():
-            a = form.cleaned_data['name']
-            b = form.cleaned_data['place']
+            if form.is_valid():
+                a = form.cleaned_data['name']
+                b = form.cleaned_data['place']
 
-            # Logica para gravar instância principal
-            b2 = Site.objects.filter(pk=pk)
-            b2.update(name=a, place=b, )
+                # Logica para gravar instância principal
+                b2 = Site.objects.filter(pk=pk)
+                b2.update(name=a, place=b, )
 
-    else: # Read data from 
-        instance = Site.objects.get(pk=pk)
-        form = SiteForm(instance=instance, prefix='edit')
-
-    context = {'form': form}
-    return render(request, 'sites/new_or_edit.html', context)
+        else: # Read data from 
+            instance = Site.objects.get(pk=pk)
+            form = SiteForm(instance=instance, prefix='edit')
+        context = {'form': form}
+        return render(request, 'sites/new_or_edit.html', context)
+    message = "You cannot have permission to do this!"
+    messages.info(request, message)
+    return home(request)
